created separate adapter for the getApartmentTemperatures
