+++ core/src/adapters/property-base-adapter/index.ts
@@ -888,36 +888,7 @@ export async function getBlockReasons(): Promise<
 
 // ==================== APARTMENT TEMPERATURES (EcoGuard Curves) ====================
 
-export async function getApartmentTemperatures(
-  objectNumber: string,
-  query: { from?: number; to?: number; interval?: 'H' | 'D' }
-): Promise<
-  AdapterResult<
-    components['schemas']['ApartmentTemperaturesResponse'],
-    'not-found' | 'unknown'
-  >
-> {
-  try {
-    const response = await client().GET(
-      '/apartments/{objectNumber}/temperatures',
-      {
-        params: { path: { objectNumber }, query },
-      }
-    )
-
-    if (response.data?.content) {
-      return { ok: true, data: response.data.content }
-    }
-    if (response.response.status === 404) {
-      return { ok: false, err: 'not-found' }
-    }
-
-    return { ok: false, err: 'unknown' }
-  } catch (err) {
-    logger.error({ err }, '@onecore/property-adapter.getApartmentTemperatures')
-    return { ok: false, err: 'unknown' }
-  }
-}
+export { getApartmentTemperatures } from './apartment-temperatures'
 
 // ==================== COMPONENTS ====================
